+++ packages/motion/template/main.js
@@ -1,8 +1,8 @@
 /* @flow */
 
-import Motion from 'react'
+import React from 'react'
 
-class Main extends Motion.Component {
+class Main extends React.Component {
   render() {
     return <div>Hello World</div>
   }