+++ packages/style/src/index.js
@@ -39,7 +39,13 @@ function Style(ComposedComponent) {
 
         // nice style them
         for (const style in styles) {
-          if (style) niceStyles(styles[style])
+          if (!styles.hasOwnProperty(style)) {
+            continue
+          }
+          const value = styles[style]
+          if (value) {
+            styles[style] = niceStyles(value)
+          }
         }
 
         this.stylesheet = StyleSheet.create(styles)