+++ scripts/__variables.sh
@@ -1,1 +1,1 @@
-PACKAGE_NAMES="babel-preset transform nice-styles motion style"
+PACKAGE_NAMES="babel-preset nice-styles motion style"